CourseManager commit

+++ src/main/java/se/lexicon/course_manager_assignment/data/service/student/StudentManager.java
@@ -80,7 +80,10 @@ public class StudentManager implements StudentService {
 
     @Override
     public boolean deleteStudent(int id) {
-        studentDao.removeStudent(studentDao.findById(id));
+        if (id != 0){
+            studentDao.removeStudent(studentDao.findById(id));
+            return true;
+        }
         return false;
     }
 }
